+++ config/config_test.go
@@ -12,7 +12,9 @@ import (
 
 	metricnoop "go.opentelemetry.io/otel/metric/noop"
 	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
+	"go.opentelemetry.io/otel/sdk/resource"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
+	semconv "go.opentelemetry.io/otel/semconv/v1.21.0"
 	tracenoop "go.opentelemetry.io/otel/trace/noop"
 )
 
@@ -49,5 +51,71 @@ func TestNewSDK(t *testing.T) {
 		assert.IsType(t, tt.wantTracerProvider, sdk.TracerProvider())
 		assert.IsType(t, tt.wantMeterProvider, sdk.MeterProvider())
 		require.Equal(t, tt.wantShutdownErr, sdk.Shutdown(context.Background()))
+	}
+}
+
+func strPtr(val string) *string {
+	return &val
+}
+
+func intPtr(val int) *int {
+	return &val
+}
+
+func TestNewResource(t *testing.T) {
+	tests := []struct {
+		name         string
+		config       *Resource
+		wantResource *resource.Resource
+	}{
+		{
+			name: "nil resource",
+		},
+		{
+			name: "resource with service name",
+			config: &Resource{
+				Attributes: &Attributes{
+					ServiceName: strPtr("myservice"),
+				},
+			},
+			wantResource: resource.NewWithAttributes(
+				semconv.SchemaURL,
+				semconv.ServiceName("myservice"),
+			),
+		},
+	}
+	for _, tt := range tests {
+		res, err := newResource(tt.config)
+		require.NoError(t, err)
+		wantedRes, err := resource.Merge(resource.Default(), tt.wantResource)
+		require.NoError(t, err)
+		require.Equal(t, wantedRes, res)
+	}
+}
+
+func TestNormalizeEndpoint(t *testing.T) {
+	tests := []struct {
+		name         string
+		endpoint     string
+		wantEndpoint string
+	}{
+		{
+			name:         "no prefix",
+			endpoint:     "localhost:1234",
+			wantEndpoint: "http://localhost:1234",
+		},
+		{
+			name:         "http prefix",
+			endpoint:     "http://localhost:1234",
+			wantEndpoint: "http://localhost:1234",
+		},
+		{
+			name:         "https prefix",
+			endpoint:     "https://localhost:1234",
+			wantEndpoint: "https://localhost:1234",
+		},
+	}
+	for _, tt := range tests {
+		require.Equal(t, tt.wantEndpoint, normalizeEndpoint(tt.endpoint))
 	}
 }